+++ Changelog.md
@@ -22,6 +22,9 @@
 - Adjustments to match the SI prefix according to IEC (Base 10) (Dendraspis)
 - Set minimum bitrate of 32kbps for (E)AC3 (Dendraspis, 44vince44)
 - Fix ffmpeg DD+ encoding issue with high bitrates by adding 0.1 to bitrate (Dendraspis, [#566](/../../issues/566))
+- Add vertical scrollbars to Code Editor code-boxes (Dendraspis, [#567](/../../issues/567))
+- Code Editor does not join Source, Crop, Resize and Rotation filters automatically (Dendraspis)
+- Code Editor can join (in-)active filters only [Don't forget to reset "Edit Menu" in Code Editor!] (Dendraspis)  
 - SMDegrain 3.1.2.111s